refactor: 수익률 계산 전략을 GameResult로 이동
