+++ src/main/java/game/GameResult.java
@@ -1,11 +1,19 @@
 package game;
 
+import java.util.function.Function;
+
 public enum GameResult {
 
-    WIN,
-    LOSE,
-    DRAW,
-    BLACKJACK;
+    WIN(betting -> betting),
+    LOSE(betting -> (-1) * betting),
+    DRAW(betting -> 0),
+    BLACKJACK(betting -> (int) (1.5 * betting));
+
+    private final Function<Integer, Integer> evaluateStrategy;
+
+    GameResult(Function<Integer, Integer> evaluateStrategy) {
+        this.evaluateStrategy = evaluateStrategy;
+    }
 
     public static GameResult of(Dealer dealer, Player player) {
         if (player.isBlackJack() && dealer.isBlackJack()) {
@@ -42,4 +50,8 @@ public enum GameResult {
         return DRAW;
     }
 
+    public int evaluate(int betting) {
+        return evaluateStrategy.apply(betting);
+    }
+    
 }
