+++ src/test/java/game/GameResultTest.java
@@ -1,5 +1,7 @@
 package game;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import card.Card;
 import card.CardNumber;
 import card.Pattern;
@@ -191,5 +193,46 @@ public class GameResultTest {
         //then
         Assertions.assertThat(gameResult).isEqualTo(GameResult.BLACKJACK);
     }
+
+    @Test
+    void 승리하면_배팅_금액을_받는다() {
+        GameResult result = GameResult.WIN;
+        Betting betting = new Betting(10000);
+
+        int profit = result.evaluate(betting.getBetting());
+
+        assertThat(profit).isEqualTo(10000);
+    }
+
+    @Test
+    void 무승부이면_수익이_없다() {
+        GameResult result = GameResult.DRAW;
+        Betting betting = new Betting(10000);
+
+        int profit = result.evaluate(betting.getBetting());
+
+        assertThat(profit).isEqualTo(0);
+    }
+
+    @Test
+    void 패배하면_배팅_금액을_잃는다() {
+        GameResult result = GameResult.LOSE;
+        Betting betting = new Betting(10000);
+
+        int profit = result.evaluate(betting.getBetting());
+
+        assertThat(profit).isEqualTo(-10000);
+    }
+
+    @Test
+    void 블랙잭이면_1_5배를_받는다() {
+        GameResult result = GameResult.BLACKJACK;
+        Betting betting = new Betting(10000);
+
+        int profit = result.evaluate(betting.getBetting());
+
+        assertThat(profit).isEqualTo(15000);
+    }
+
 
 }
